requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==4.2.16
django-crispy-forms==2.3
sqlparse==0.5.2
typing-extensions==4.12.2
tzdata==2024.2
gunicorn==21.2.0
Pillow==11.1.0
whitenoise==6.6.0
psycopg2==2.9.10
python-dotenv==1.0.1
dj-database-url==2.1.0
cloudinary==1.42.1
django-cloudinary-storage==0.3.0
django-countries==7.6.1